requirements.txt
alembic==1.0.3
altgraph==0.16.1
alyn==0.1.1
aniso8601==4.0.1
appdirs==1.4.3
argcomplete==1.8.2
arrow==0.13.0
atomicwrites==1.3.0
attrs==18.2.0
auto-py-to-exe==2.5.1
autoenv==1.0.0
beautifulsoup4==4.5.3
blinker==1.4
bokeh==1.0.4
bottle==0.12.16
bottle-websocket==0.2.9
cachetools==2.1.0
certifi==2018.8.13
chardet==3.0.4
Click==7.0
clickclick==1.2.2
cloudpickle==0.6.1
commonregex==1.5.4
connexion==2.0.2
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dask==0.20.2
date-extractor==3.9.1
datefinder==0.7.0
dateparser==0.7.0
decorator==4.3.0
dill==0.2.8.2
distro==1.3.0
Django==2.1.3
dlib==19.16.0
docopt==0.6.2
docx2txt==0.6
EbookLib==0.15
Eel==0.9.11
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
Flask-Cors==3.0.7
Flask-DotEnv==0.1.1
Flask-Environments==0.1
Flask-Exceptions==1.2.1
Flask-JWT-Extended==3.13.1
Flask-Mail==0.9.0
flask-marshmallow==0.9.0
Flask-Migrate==2.3.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
fpdf==1.7.2
future==0.17.1
gevent==1.4.0
gevent-websocket==0.10.1
google==2.0.1
google-api-core==1.3.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-language==1.1.1
google-cloud-vision==0.33.0
googleapis-common-protos==1.5.3
greenlet==0.4.15
grpcio==1.14.1
gunicorn==19.9.0
httplib2==0.11.3
hurry==1.0
idna==2.7
image==1.5.27
imageio==2.4.1
imdirect==0.5.0
img-rotate==0.1.1
imutils==0.5.2
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
kiwisolver==1.0.1
lxml==4.2.6
macholib==1.11
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.16.3
marshmallow-sqlalchemy==0.15.0
matplotlib==3.0.2
mccabe==0.6.1
moment==0.8.2
more-itertools==5.0.0
mrz==0.3.5
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
mysqlclient==1.3.13
nameparser==1.0.2
networkx==2.2
nltk==3.4
numpy==1.15.4
openapi-spec-validator==0.2.4
opencv-contrib-python==4.0.0.21
opencv-python==3.4.4.19
packaging==19.0
pandas==0.23.4
passlib==1.7.1
PassportEye==1.4.0
pathlib2==2.3.3
pdfminer.six==20181108
pdfminer2==20151206
pdfminer3k==1.3.1
pefile==2018.8.8
piexif==1.1.2
Pillow==5.3.0
pkg-resources==0.0.0
plac==0.9.6
pluggy==0.8.1
ply==3.11
pocketsphinx==0.1.3
preshed==2.0.1
protobuf==3.6.1
protobuf3-to-dict==0.1.5
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.5.0
pycountry==18.12.8
pycryptodome==3.7.2
pyflakes==2.1.1
PyInstaller==3.4
PyJWT==1.6.4
pyocr==0.5.3
pyparsing==2.3.0
PyPDF2==1.26.0
pypng==0.0.18
PyQRCode==1.2.1
pytesseract==0.2.5
pytest==4.2.0
python-dateutil==2.7.5
python-editor==1.0.3
python-pptx==0.6.5
pytz==2018.5
PyWavelets==1.0.1
PyYAML==3.13
pyzbar==0.1.7
qrtools==0.0.2
regex==2018.1.10
requests==2.19.1
rsa==3.4.2
schedule==0.6.0
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
scrapy-corenlp==0.2.0
simple-http-server==0.1.7
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
slate==0.3
sortedcontainers==2.1.0
spacy==2.0.18
SpeechRecognition==3.6.3
SQLAlchemy==1.2.14
tabula==1.0.5
tabula-py==1.3.1
textract==1.6.1
thinc==6.12.1
times==0.7
tinify==1.5.1
toolz==0.9.0
tornado==6.0.2
tqdm==4.28.1
typing==3.6.6
tzlocal==1.5.1
ujson==1.35
uritemplate==3.0.0
uritools==2.2.0
urlextract==0.9
urllib3==1.23
Wand==0.5.1
Werkzeug==0.14.1
whichcraft==0.5.2
wrapt==1.10.11
xlrd==1.0.0
XlsxWriter==1.1.2
xmltodict==0.11.0
zbarlight==2.1
